Added code in assignment, student (incomplete and just description) course was not modified added and removed simply, department added letter before  Id (was only numbers before)

+++ src/main/java/org/example/Course.java
@@ -25,7 +25,7 @@ public class Course {
     public boolean isAssignmentWeightValid(){
         int totalWeight = 0;
         for (Assignment assigment : assignments) {
-            totalWeight += assigment.getWeight();
+            totalWeight += (int) assigment.getWeight();
             if (totalWeight == 100) {
                 return true;
             }
